+++ R/methods__token.R
@@ -867,7 +867,13 @@ refresh_token_dispatch <- function(
   S7::check_is_S7(token, OAuthToken)
   if (token_targets_configured(oauth_client) && is.null(target_request)) {
     limits <- token_target_limits(oauth_client)
-    limits[[oauth_client@default_token_target]] <- token@granted_scopes
+    limits[[
+      oauth_client@default_token_target
+    ]] <- token_target_operation_scopes(
+      oauth_client,
+      oauth_client@default_token_target,
+      token@granted_scopes
+    )
     target_request <- token_target_request(
       oauth_client,
       limits = limits,